backend/requirements.txt
flask==2.3.3
flask-socketio==5.3.6
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
yfinance==0.2.18
sqlite3
tensorflow==2.15.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4